update Silverado ff fit
this is also for Sierra, which fingerprints as a Silverado. I used a combined data set for both trucks, and data from the Silverado was insufficient so it's weighted towards the Sierra. Regardless, it fits the data of both very well.

For the combined data set where "old" fit is the stock feedforward it was using before the torque controller:
MAE old 0.2519, new 0.0504
STD old 0.1208, new 0.0516

now a fit for just the Sierra data, where the "old" fit is the combined fit:
MAE old 0.0548, new 0.0524
STD old 0.0636, new 0.0628

and the Silverado fit vs combined "old" fit:
MAE old 0.0538, new 0.0473
STD old 0.0394, new 0.0378

the Silverado benefits a little from using only its data, but that data was incomplete anyhow.